renamed files corrected spelling
