+++ frontend/pantry/src/components/Receipts/ReceiptList.tsx
