+++ frontend/pantry/src/components/Receipts/Receipts.tsx
@@ -10,13 +10,13 @@ import {
 import styles from './styles.ts';
 
 import dummyStores from './dummyData/dummyStores.ts';
-import dummyReceipts from './dummyData/dummyReciepts.ts';
-import RecieptList from './RecieptList.tsx';
+import dummyReceipts from './dummyData/dummyReceipts.ts';
+import RecieptList from './ReceiptList.tsx';
 import StorePicker from './StorePicker.tsx';
 
-interface Props {}
+// interface Props {}
 
-const Reciepts = (props: Props) => {
+const Reciepts = (props: any) => {
   const [selectedValue, setSelectedValue] = useState('-');
   const handleStoreChange = (itemValue: string, itemIndex: number) => {
     console.log(itemValue);
